requirements.txt
torch
torchvision
flask
datasets
numpy
pillow
scikit-learn
